remove confirm email  🐿 v2.12.4
